+++ QTap-android/app/src/main/java/com/example/alex/qtapandroid/common/database/food/FoodManager.java
@@ -140,11 +140,11 @@ public class FoodManager extends DatabaseAccessor {
             //getInt()>0 because SQLite doesn't have boolean types - 1 is true, 0 is false
             food = new Food(cursor.getString(Food.POS_NAME), cursor.getInt(Food.POS_BUILDING_ID),
                     cursor.getString(Food.POS_INFORMATION), cursor.getInt(Food.POS_MEAL_PLAN) > 0, cursor.getInt(Food.POS_CARD) > 0,
-                    cursor.getFloat(Food.POS_MON_START_HOURS), cursor.getFloat(Food.POS_MON_STOP_HOURS), cursor.getFloat(Food.POS_TUE_START_HOURS),
-                    cursor.getFloat(Food.POS_TUE_STOP_HOURS), cursor.getFloat(Food.POS_WED_START_HOURS), cursor.getFloat(Food.POS_WED_STOP_HOURS),
-                    cursor.getFloat(Food.POS_THUR_START_HOURS), cursor.getFloat(Food.POS_THUR_STOP_HOURS), cursor.getFloat(Food.POS_FRI_START_HOURS),
-                    cursor.getFloat(Food.POS_FRI_STOP_HOURS), cursor.getFloat(Food.POS_SAT_START_HOURS),
-                    cursor.getFloat(Food.POS_SAT_STOP_HOURS), cursor.getFloat(Food.POS_SUN_START_HOURS), cursor.getFloat(Food.POS_SUN_STOP_HOURS));
+                    cursor.getDouble(Food.POS_MON_START_HOURS), cursor.getDouble(Food.POS_MON_STOP_HOURS), cursor.getDouble(Food.POS_TUE_START_HOURS),
+                    cursor.getDouble(Food.POS_TUE_STOP_HOURS), cursor.getDouble(Food.POS_WED_START_HOURS), cursor.getDouble(Food.POS_WED_STOP_HOURS),
+                    cursor.getDouble(Food.POS_THUR_START_HOURS), cursor.getDouble(Food.POS_THUR_STOP_HOURS), cursor.getDouble(Food.POS_FRI_START_HOURS),
+                    cursor.getDouble(Food.POS_FRI_STOP_HOURS), cursor.getDouble(Food.POS_SAT_START_HOURS),
+                    cursor.getDouble(Food.POS_SAT_STOP_HOURS), cursor.getDouble(Food.POS_SUN_START_HOURS), cursor.getDouble(Food.POS_SUN_STOP_HOURS));
             food.setID(cursor.getInt(Food.POS_ID));
             cursor.close();
             return food; //return only when the cursor has been closed.
@@ -190,11 +190,11 @@ public class FoodManager extends DatabaseAccessor {
             //getInt()>0 because SQLite doesn't have boolean types - 1 is true, 0 is false
             food = new Food(cursor.getString(Food.POS_NAME), cursor.getInt(Food.POS_BUILDING_ID),
                     cursor.getString(Food.POS_INFORMATION), cursor.getInt(Food.POS_MEAL_PLAN) > 0, cursor.getInt(Food.POS_CARD) > 0,
-                    cursor.getFloat(Food.POS_MON_START_HOURS), cursor.getFloat(Food.POS_MON_STOP_HOURS), cursor.getFloat(Food.POS_TUE_START_HOURS),
-                    cursor.getFloat(Food.POS_TUE_STOP_HOURS), cursor.getFloat(Food.POS_WED_START_HOURS), cursor.getFloat(Food.POS_WED_STOP_HOURS),
-                    cursor.getFloat(Food.POS_THUR_START_HOURS), cursor.getFloat(Food.POS_THUR_STOP_HOURS), cursor.getFloat(Food.POS_FRI_START_HOURS),
-                    cursor.getFloat(Food.POS_FRI_STOP_HOURS), cursor.getFloat(Food.POS_SAT_START_HOURS),
-                    cursor.getFloat(Food.POS_SAT_STOP_HOURS), cursor.getFloat(Food.POS_SUN_START_HOURS), cursor.getFloat(Food.POS_SUN_STOP_HOURS));
+                    cursor.getDouble(Food.POS_MON_START_HOURS), cursor.getDouble(Food.POS_MON_STOP_HOURS), cursor.getDouble(Food.POS_TUE_START_HOURS),
+                    cursor.getDouble(Food.POS_TUE_STOP_HOURS), cursor.getDouble(Food.POS_WED_START_HOURS), cursor.getDouble(Food.POS_WED_STOP_HOURS),
+                    cursor.getDouble(Food.POS_THUR_START_HOURS), cursor.getDouble(Food.POS_THUR_STOP_HOURS), cursor.getDouble(Food.POS_FRI_START_HOURS),
+                    cursor.getDouble(Food.POS_FRI_STOP_HOURS), cursor.getDouble(Food.POS_SAT_START_HOURS),
+                    cursor.getDouble(Food.POS_SAT_STOP_HOURS), cursor.getDouble(Food.POS_SUN_START_HOURS), cursor.getDouble(Food.POS_SUN_STOP_HOURS));
             food.setID(cursor.getInt(Food.POS_ID));
             cursor.close();
             return food; //return only when the cursor has been closed.
